move factories to tools.llm
